Resolving minor issues and cleaning up tests
