+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/.classpath
@@ -5,7 +5,7 @@
 	<classpathentry kind="src" path="test-bundles/ffdc-bundle/src"/>
 	<classpathentry kind="src" path="test-applications/MpTelemetryLogApp/src"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-11">
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
 		<attributes>
 			<attribute name="module" value="true"/>
 		</attributes>
