+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/fat/src/io/openliberty/microprofile/telemetry/logging/internal_fat/TelemetryFFDCTest.java
@@ -92,6 +92,7 @@ public class TelemetryFFDCTest extends FATServletClient {
 
     static void testTelemetryFFDCMessages(LibertyServer s, Consumer<List<String>> consoleConsumer) throws Exception {
         hitWebPage(s, "ffdc-servlet", "FFDCServlet", true, "?generateFFDC=true");
+        Thread.sleep(2000);
 
         String logLevelLine = s.waitForStringInLog(".*scopeInfo.*FFDC_TEST_DOGET", s.getConsoleLogFile());
         String exceptionMessageLine = s.waitForStringInLog("exception.message=\"FFDC_TEST_DOGET\"", s.getConsoleLogFile());
