+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/publish/files/invalidSource.xml
@@ -20,5 +20,4 @@
 	<include location="../fatTestPorts.xml"/>
 	<mpTelemetry source="msg"/>
 	<application location="MpTelemetryLogApp.war"/>
-
 </server>
