+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/publish/servers/TelemetryDropins/server.env
@@ -1,1 +1,2 @@
 OTEL_METRICS_EXPORTER=none
+OTEL_TRACE_EXPORTER=none
