+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/publish/servers/TelemetryMessage/server.env
@@ -1,3 +1,4 @@
 OTEL_LOGS_EXPORTER=logging
 OTEL_METRICS_EXPORTER=none
+OTEL_TRACE_EXPORTER=none
 OTEL_SDK_DISABLED=false
