+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/publish/servers/TelemetrySources/server.xml
@@ -11,7 +11,6 @@
         IBM Corporation - initial API and implementation
  -->
 <server description="Telemetry Server to test MPTelemetry sources attribute">
-	<logging traceSpecification="io.openliberty.microprofile.telemetry.logging.internal.fat.logstashapp.*=all" />
 	<featureManager>
 		<feature>mpTelemetry-2.0</feature>
 		<feature>componentTest-2.0</feature>
