+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/test-applications/MpTelemetryLogApp/src/io/openliberty/microprofile/telemetry/logging/internal/fat/MpTelemetryLogApp/ExtensionServlet.java
@@ -1,3 +1,15 @@
+/*******************************************************************************
+ * Copyright (c) 2024 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
 package io.openliberty.microprofile.telemetry.logging.internal.fat.MpTelemetryLogApp;
 
 import java.io.IOException;
