+++ dev/io.openliberty.microprofile.telemetry.2.0.logging.internal_fat/test-applications/MpTelemetryLogApp/src/io/openliberty/microprofile/telemetry/logging/internal/fat/MpTelemetryLogApp/LogServlet.java
@@ -1,6 +1,5 @@
-
 /*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
+ * Copyright (c) 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
